document loading working

+++ src/app/db/loaders.ts
@@ -1,4 +1,4 @@
-import { Document } from './data'
+import { Document, Author } from './data'
 
 export const loadDocument = (id: string): Promise<Document> => {
   return new Promise((resolve, reject) => {
@@ -43,6 +43,35 @@ export const loadAllDocuments = (): Promise<{ [key: string]: Document }> => {
       )
 
       worker.onmessage = (e: MessageEvent<{ [key: string]: Document }>) => {
+        resolve(e.data)
+        worker.terminate()
+      }
+
+      worker.onerror = (error) => {
+        reject(error)
+        worker.terminate()
+      }
+
+      worker.postMessage('LOAD')
+    } else {
+      reject(
+        new Error(
+          'Web Workers are not supported in this environment. Please avoid using a prehistoric browser.'
+        )
+      )
+    }
+  })
+}
+
+export const loadAllAuthors = (): Promise<{ [key: string]: Author }> => {
+  return new Promise((resolve, reject) => {
+    if (typeof Worker !== 'undefined') {
+      const worker = new Worker(
+        new URL('./workers/documentLoader.worker.ts', import.meta.url),
+        { type: 'module' }
+      )
+
+      worker.onmessage = (e: MessageEvent<{ [key: string]: Author }>) => {
         resolve(e.data)
         worker.terminate()
       }
